Added field mapper

+++ js/admin.js
@@ -12,18 +12,21 @@ jQuery(document).ready(function($) {
 		jQuery('.wpcf7-pdf-forms-admin .messages').append(
 			jQuery('<div class="error"/>').text(msg)
 		);
+		location.href = '#wpcf7-pdf-form-messages';
 	};
 	
 	var warningMessage = function(msg) {
 		jQuery('.wpcf7-pdf-forms-admin .messages').append(
 			jQuery('<div class="warning"/>').text(msg)
 		);
+		location.href = '#wpcf7-pdf-form-messages';
 	};
 	
 	var successMessage = function(msg) {
 		jQuery('.wpcf7-pdf-forms-admin .messages').append(
 			jQuery('<div class="updated"/>').text(msg)
 		);
+		location.href = '#wpcf7-pdf-form-messages';
 	};
 	
 	var showSpinner = function() {
@@ -40,7 +43,7 @@ jQuery(document).ready(function($) {
 		;
 	}
 	
-	var getTags = function(attachment_id) {
+	var getTags = function(attachments, all = false) {
 		
 		clearMessages();
 		
@@ -51,7 +54,7 @@ jQuery(document).ready(function($) {
 		jQuery.ajax({
 			url: wpcf7_pdf_forms.ajax_url,
 			type: 'GET',
-			data: { 'action': 'wpcf7_pdf_forms_query_fields', 'attachment_id': attachment_id, 'nonce': wpcf7_pdf_forms.ajax_nonce },
+			data: { 'action': 'wpcf7_pdf_forms_query_tags', 'attachments': attachments, 'all': all, 'nonce': wpcf7_pdf_forms.ajax_nonce },
 			cache: false,
 			dataType: 'json',
 			
@@ -61,7 +64,10 @@ jQuery(document).ready(function($) {
 					return errorMessage(data.error_message);
 				
 				if(data.tags)
+				{
 					textarea.val(data.tags);
+					location.href = '#wpcf7-pdf-form-tags-textarea';
+				}
 			},
 			
 			error: function (jqXHR, textStatus, errorThrown) { return errorMessage(textStatus); },
@@ -72,29 +78,194 @@ jQuery(document).ready(function($) {
 		});
 	};
 	
-	var getAttachments = function() {
-		var value = jQuery('textarea#wpcf7-form').closest('form').find('input[name=wpcf7-pdf-forms-attachments]').val();
-		if(value)
-			return JSON.parse(value);
-		else
-			return [];
+	var pdfFieldsCache = [];
+	var reloadPdfFieldsCounter = 0;
+	var reloadPdfFieldsCounterXhr;
+	var reloadPdfFields = function() {
+		
+		var attachments = getAttachments();
+		var attachmentList = [];
+		
+		for(var i=0, l=attachments.length; i<l; i++)
+			attachmentList.push(attachments[i].attachment_id);
+		
+		var reloadIndex = ++reloadPdfFieldsCounter;
+		if(reloadPdfFieldsCounterXhr)
+			reloadPdfFieldsCounterXhr.abort();
+		
+		reloadPdfFieldsCounterXhr = jQuery.ajax({
+			url: wpcf7_pdf_forms.ajax_url,
+			type: 'GET',
+			data: { 'action': 'wpcf7_pdf_forms_query_pdf_fields', 'attachments': attachmentList, 'nonce': wpcf7_pdf_forms.ajax_nonce },
+			cache: false,
+			dataType: 'json',
+			
+			success: function (data, textStatus, jqXHR) {
+				
+				reloadPdfFieldsCounterXhr = null;
+				
+				if(!data.success)
+					return errorMessage(data.error_message);
+				
+				if(data.fields)
+				{
+					if(reloadIndex!=reloadPdfFieldsCounter)
+						return;
+					
+					pdfFieldsCache = data.fields;
+					
+					preloadMappings();
+					refreshPdfFields();
+				}
+			},
+			
+			error: function (jqXHR, textStatus, errorThrown) { if(!jqXHR.getAllResponseHeaders()) return; return errorMessage(textStatus); },
+			
+			beforeSend: function() { showSpinner() },
+			complete: function() { hideSpinner(); }
+			
+		});
+	};
+	var getPdfFieldData = function(id) {
+		
+		for(var i=0, l=pdfFieldsCache.length; i<l; i++)
+			if(pdfFieldsCache[i].id == id)
+				return pdfFieldsCache[i]
+		
+		return null;
+	};
+	
+	var refreshPdfFields = function() {
+		
+		var pdf_fields = jQuery('.wpcf7-pdf-forms-admin .pdf-field-list');
+		pdf_fields.empty();
+		
+		var mappings = getMappings();
+		
+		jQuery.each(pdfFieldsCache, function (f, field) {
+			
+			for(var i=0, l=mappings.length; i<l; i++)
+			{
+				mapped_id = String(mappings[i].pdf_field);
+				field_id = String(field.id);
+				mapped_id = mapped_id.substr(mapped_id.indexOf('-')+1);
+				field_id = field_id.substr(field_id.indexOf('-')+1);
+				if(mapped_id == field_id)
+					return;
+			}
+			
+			pdf_fields.append(jQuery('<option>', { 
+				value: field.id,
+				text : field.caption
+			}));
+		});
+		
+		updateTagHint();
+	};
+	
+	var cf7FieldsCache = [];
+	var loadCf7Fields = function(callback = null) {
+		
+		var cf7_fields = jQuery('.wpcf7-pdf-forms-admin .cf7-field-list');
+		
+		var form = jQuery('textarea#wpcf7-form');
+		
+		jQuery.ajax({
+			url: wpcf7_pdf_forms.ajax_url,
+			type: 'GET',
+			data: { 'action': 'wpcf7_pdf_forms_query_cf7_fields', 'wpcf7-form': form.val(), 'nonce': wpcf7_pdf_forms.ajax_nonce },
+			cache: false,
+			dataType: 'json',
+			
+			success: function (data, textStatus, jqXHR) {
+				
+				if(!data.success)
+					return errorMessage(data.error_message);
+				
+				if(data.fields)
+				{
+					cf7FieldsCache = data.fields;
+					
+					cf7_fields.empty();
+					
+					jQuery.each(data.fields, function (i, field) {
+						
+						cf7_fields.append(jQuery('<option>', {
+							value: field.id,
+							text : field.caption 
+						}));
+					});
+					
+					if(callback)
+						callback();
+				}
+			},
+			
+			error: function (jqXHR, textStatus, errorThrown) { return errorMessage(textStatus); },
+			
+			beforeSend: function() { showSpinner() },
+			complete: function() { hideSpinner(); }
+			
+		});
+	};
+	var getCf7FieldData = function(id) {
+		
+		for(var i=0, l=cf7FieldsCache.length; i<l; i++)
+			if(cf7FieldsCache[i].id == id)
+				return cf7FieldsCache[i]
+		
+		return null;
+	};
+	var addDefaultMappings = function() {
+		jQuery.each(cf7FieldsCache, function (i, field) {
+			if(field.pdf_field)
+				addMapping(field.id, field.pdf_field);
+		});
 	}
 	
-	var setAttachments = function(attachments) {
+	var getData = function(field) {
+		var data = jQuery('textarea#wpcf7-form').closest('form').find('input[name=wpcf7-pdf-forms-data]').val();
+		if(data)
+			data = JSON.parse(data);
+		else
+			data = {};
+		return data[field];
+	};
+	
+	var setData = function(field, value) {
 		jQuery('textarea#wpcf7-form').each(function() {
 			var form = jQuery('textarea#wpcf7-form').closest('form');
 			if(form[0])
 			{
-				var input = form.find('input[name=wpcf7-pdf-forms-attachments]');
+				var input = form.find('input[name=wpcf7-pdf-forms-data]');
 				if(!input[0])
 				{
-					input = jQuery("<input type='hidden' name='wpcf7-pdf-forms-attachments'/>");
+					input = jQuery("<input type='hidden' name='wpcf7-pdf-forms-data'/>");
 					jQuery(form).append(input);
 				}
-				input.val(JSON.stringify(attachments));
+				var data = input.val();
+				if(data)
+					data = JSON.parse(data);
+				else
+					data = {};
+				data[field] = value;
+				input.val(JSON.stringify(data));
 			}
 		});
-	}
+	};
+	
+	var getAttachments = function() {
+		var attachments = getData('attachments');
+		if(attachments)
+			return attachments;
+		else
+			return [];
+	};
+	
+	var setAttachments = function(attachments) {
+		setData('attachments', attachments);
+		reloadPdfFields();
+	};
 	
 	var deleteAttachment = function(attachment_id) {
 		
@@ -135,9 +306,11 @@ jQuery(document).ready(function($) {
 		
 		jQuery('.wpcf7-pdf-forms-admin .instructions').remove();
 		
-		var template = jQuery('.pdf-attachment-row-template');
+		var template = jQuery('.wpcf7-pdf-forms-admin .pdf-attachment-row-template');
 		var tag = template.clone().removeClass('pdf-attachment-row-template');
+		
 		tag.find('.filename').text('['+attachment_id+'] '+filename);
+		
 		if(typeof options != 'undefined' && options !== null)
 		{
 			tag.find('.pdf-options input[type=checkbox]').each(function(){
@@ -163,9 +336,10 @@ jQuery(document).ready(function($) {
 			event.stopPropagation();
 			event.preventDefault();
 			
-			getTags(jQuery(this).data('attachment_id'));
+			getTags([jQuery(this).data('attachment_id')]);
 			
 		});
+		
 		var delete_button = tag.find('.delete-button');
 		delete_button.data('attachment_id', attachment_id);
 		delete_button.click(function(event) {
@@ -174,16 +348,18 @@ jQuery(document).ready(function($) {
 			event.stopPropagation();
 			event.preventDefault();
 			
+			if(!confirm(wpcf7_pdf_forms.__Confirm_Delete_Attachment))
+				return;
+			
 			deleteAttachment(jQuery(this).data('attachment_id'));
 			
 			tag.remove();
 		});
+		
 		jQuery('.wpcf7-pdf-forms-admin .pdf-attachments').append(tag);
 	};
 	
 	var preloadAttachments = function() {
-		
-		clearMessages();
 		
 		if(!post_id)
 			return;
@@ -212,9 +388,174 @@ jQuery(document).ready(function($) {
 		});
 	};
 	
-	// set up 'Insert Tags' button handler
-	jQuery('.wpcf7-pdf-forms-admin .insert-tags-btn').click(function(event) {
-		var tags = jQuery('.wpcf7-pdf-forms-admin .tags-textarea').val();
+	var getMappings = function() {
+		var mappings = getData('mappings');
+		if(mappings)
+			return mappings;
+		else
+			return [];
+	};
+	
+	var setMappings = function(mappings) {
+		setData('mappings', mappings);
+		refreshPdfFields();
+	};
+	
+	var deleteMapping = function(cf7_field, pdf_field) {
+		
+		var mappings = getMappings();
+		
+		for(var i=0, l=mappings.length; i<l; i++)
+			if(mappings[i].cf7_field == cf7_field
+			&& mappings[i].pdf_field == pdf_field)
+			{
+				mappings.splice(i, 1);
+				break;
+			}
+		
+		setMappings(mappings);
+	};
+	
+	var addMapping = function(cf7_field, pdf_field) {
+		
+		var cf7_field_data = getCf7FieldData(cf7_field);
+		if(!cf7_field_data)
+			return;
+		var pdf_field_data = getPdfFieldData(pdf_field);
+		if(!pdf_field_data)
+			return;
+		
+		var virtual = cf7_field_data.pdf_field == pdf_field;
+		
+		var mappings = getMappings();
+		mappings.push( { 'cf7_field': cf7_field, 'pdf_field': pdf_field } );
+		setMappings(mappings);
+		
+		var template = jQuery('.wpcf7-pdf-forms-admin .pdf-mapping-row-template');
+		var tag = template.clone().removeClass('pdf-mapping-row-template');
+		
+		tag.find('.cf7-field-name').text(cf7_field_data.caption);
+		tag.find('.pdf-field-name').text(pdf_field_data.caption);
+		
+		var delete_button = tag.find('.delete-mapping-button');
+		if(virtual)
+			delete_button.remove();
+		else
+		{
+			delete_button.data('cf7_field', cf7_field);
+			delete_button.data('pdf_field', pdf_field);
+			delete_button.click(function(event) {
+				
+				// prevent running default button click handlers
+				event.stopPropagation();
+				event.preventDefault();
+				
+				if(!confirm(wpcf7_pdf_forms.__Confirm_Delete_Mapping))
+					return;
+				
+				deleteMapping(jQuery(this).data('cf7_field'), jQuery(this).data('pdf_field'));
+				
+				tag.remove();
+			});
+		}
+		
+		jQuery('.wpcf7-pdf-forms-admin .pdf-fields-mapper').append(tag);
+	};
+	
+	var loadedMappings = false;
+	var preloadMappings = function() {
+		
+		if(!post_id)
+			return;
+		
+		if(loadedMappings)
+			return;
+		
+		loadedMappings = true;
+		
+		jQuery.ajax({
+			url: wpcf7_pdf_forms.ajax_url,
+			type: 'GET',
+			data: { 'action': 'wpcf7_pdf_forms_query_mappings', 'post_id': post_id, 'nonce': wpcf7_pdf_forms.ajax_nonce },
+			cache: false,
+			dataType: 'json',
+			
+			success: function (data, textStatus, jqXHR) {
+				
+				if(!data.success)
+					return errorMessage(data.error_message);
+				
+				if(data.mappings)
+					jQuery.each(data.mappings, function(index, data) { addMapping(data.cf7_field, data.pdf_field); });
+				
+				addDefaultMappings();
+			},
+			
+			error: function (jqXHR, textStatus, errorThrown) { return errorMessage(textStatus); },
+			
+			beforeSend: function() { showSpinner() },
+			complete: function() { hideSpinner(); }
+			
+		});
+	};
+	
+	var updateTagHint = function() {
+		
+		jQuery('.wpcf7-pdf-forms-admin .tag-hint').text('');
+		
+		var pdf_field = jQuery('.wpcf7-pdf-forms-admin .pdf-field-list').val();
+		if(!pdf_field)
+			return;
+		
+		var pdf_field_data = getPdfFieldData(pdf_field);
+		if(!pdf_field_data)
+			return;
+		
+		jQuery.ajax({
+			url: wpcf7_pdf_forms.ajax_url,
+			type: 'GET',
+			data: { 'action': 'wpcf7_pdf_forms_query_tag_hint', 'attachment_id': pdf_field_data.attachment_id, 'field': pdf_field_data.name, 'nonce': wpcf7_pdf_forms.ajax_nonce },
+			cache: false,
+			dataType: 'json',
+			
+			success: function (data, textStatus, jqXHR) {
+				
+				if(!data.success)
+					return errorMessage(data.error_message);
+				
+				if(data.tag_hint)
+					jQuery('.wpcf7-pdf-forms-admin .tag-hint').text(data.tag_hint);
+			},
+			
+			error: function (jqXHR, textStatus, errorThrown) { clearMessages(); return errorMessage(textStatus); },
+			
+			beforeSend: function() { showSpinner() },
+			complete: function() { hideSpinner(); }
+			
+		});
+	};
+	jQuery('.wpcf7-pdf-forms-admin .pdf-field-list').change(updateTagHint);
+	
+	// set up global 'Get Tags' button handler
+	jQuery('.wpcf7-pdf-forms-admin .get-tags-all-button').click(function(event) {
+		
+		// prevent running default button click handlers
+		event.stopPropagation();
+		event.preventDefault();
+		
+		var attachments = getAttachments();
+		if(attachments.length == 0)
+			return;
+		
+		var attachmentList = [];
+		
+		for(var i=0, l=attachments.length; i<l; i++)
+			attachmentList.push(attachments[i].attachment_id);
+		
+		getTags(attachmentList, true);
+	});
+	
+	var getWpcf7obj = function() {
 		var wpcf7obj;
 		if(typeof _wpcf7 != 'undefined' && _wpcf7 !== null)
 			wpcf7obj = _wpcf7;
@@ -222,9 +563,28 @@ jQuery(document).ready(function($) {
 			wpcf7obj = wpcf7;
 		if(!wpcf7obj)
 			alert(wpcf7_pdf_forms.__No_WPCF7);
-		else
+		return wpcf7obj;
+	};
+	
+	// set up 'Insert Tags' button handler
+	jQuery('.wpcf7-pdf-forms-admin .insert-tags-btn').click(function(event) {
+		var tags = jQuery('.wpcf7-pdf-forms-admin .tags-textarea').val();
+		var wpcf7obj = getWpcf7obj();
+		if(wpcf7obj)
 		{
 			wpcf7obj.taggen.insert(tags);
+			tb_remove();
+		}
+		return false;
+	});
+	
+	// set up 'Insert Tag' button handler
+	jQuery('.wpcf7-pdf-forms-admin .insert-tag-hint-btn').click(function(event) {
+		var tag = jQuery('.wpcf7-pdf-forms-admin .tag-hint').text();
+		var wpcf7obj = getWpcf7obj();
+		if(wpcf7obj)
+		{
+			wpcf7obj.taggen.insert(tag);
 			tb_remove();
 		}
 		return false;
@@ -274,7 +634,7 @@ jQuery(document).ready(function($) {
 				file.val("");
 				
 				if(data.attachment_id)
-					setTimeout( function(){ getTags(data.attachment_id); }, 100);
+					setTimeout( function(){ getTags([data.attachment_id]); }, 100);
 			},
 			
 			error: function (jqXHR, textStatus, errorThrown) { return errorMessage(textStatus); },
@@ -285,5 +645,28 @@ jQuery(document).ready(function($) {
 		});
 	});
 	
-	preloadAttachments();
+	
+	// set up 'Add Mapping' button handler
+	jQuery('.wpcf7-pdf-forms-admin .add-mapping-button').click(function(event) {
+		
+		// prevent running default button click handlers
+		event.stopPropagation();
+		event.preventDefault();
+		
+		clearMessages();
+		
+		var cf7_field = jQuery('.wpcf7-pdf-forms-admin .cf7-field-list').val();
+		var pdf_field = jQuery('.wpcf7-pdf-forms-admin .pdf-field-list').val();
+		
+		if(cf7_field && pdf_field)
+			addMapping(cf7_field, pdf_field);
+	});
+	
+	
+	var form = jQuery('textarea#wpcf7-form');
+	form.change(function(){
+		loadCf7Fields();
+	});
+	
+	loadCf7Fields(preloadAttachments);
 });
